+++ src/lib/constants/packageInfo.ts
@@ -1,0 +1,3 @@
+import packageInfo from "../../../package.json";
+
+export { packageInfo };